HY_ver/dev/OneTrip_dev/backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
redis>=5.0.0
supabase>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Processing
pydantic>=2.4.0
pydantic-settings>=2.0.0

# File Processing
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.13
python-dateutil>=2.8.0

# NLP and AI (simplified for initial setup)
openai>=1.3.0
requests>=2.31.0
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Monitoring
structlog>=23.1.0

# Utils
email-validator>=2.0.0